Adds certifications for Original firmware for Oryx Pro 11 under both Pop 22.04 and Ubuntu 22.04 with release e5c3632.
Notes about release: On-going issues involving the behaviour of
external monitors on boot are present in this release. The modes
of use where this behaviour manifests - namely booting with the
lid closed and an external monitor attached - is uncommon, and at
time of release is not considered to represent a blocking
behaviour. Future updates should explicitely test for resolution
of these issues.